Fixed an oversight in signal mapping
To properly map the signal names a check on the original tile type needs to be done. For this the tile type is added to the tables and a separate table is made for getting to the tile types based on the coordinates. The xoff and yoff in the signal name mapping tables are used to get to the original tile.